+++ CHANGES.rst
@@ -8,6 +8,9 @@ Unreleased (2.9.0) (XXXX-XX-XX)
 
 - Favourite: Replace state when loading favourite using mergeWith.
 
+- Fix bug in image carousel, showing the same image twice when changing time
+  interval.
+
 
 Release 2.10.4 (2016-3-22)
 ---------------------