+++ CHANGES.rst
@@ -25,6 +25,10 @@ Release 2.10.4 (2016-3-22)
 
 - JSHint prettify.
 
+- Fix bug #1485. Elevation / line data is now downloadable again.
+
+- Fix bug #1555 of nens/lizard-nxt. Timeseries can yet again export to csv.
+
 - Remove timeseries of removed assets.
 
 - Search box has a starting query of ""