+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog of lizard-nxt client
 Unreleased (2.9.0) (XXXX-XX-XX)
 -------------------------------
 
+- Fix #1596: bulb hover for profile line is back.
+
 - Favourite: Fix #1578 restore area and geometries in favourite.
 
 - Favourite: Replace state when loading favourite using mergeWith.