+++ CHANGES.rst
@@ -3,6 +3,11 @@ Changelog of lizard-nxt client
 
 Unreleased (2.9.0) (XXXX-XX-XX)
 -------------------------------
+-
+
+
+Release 2.11.1 (2016-3-25)
+---------------------
 
 - Fix #1596: bulb hover for profile line is back.